refactor(dashboard): extract Impact sections into shared insights/ module
Pull the cost / cycle-time / engagement sections out of InsightsClient.tsx so
the Analytics page (Phase 3) can render them too. No behavior change.

New packages/dashboard/components/insights/:
- types.ts        — shared InstallationFPInsight types (Insight, Cost,
                     CycleTime, Engagement, Percentiles, CategoryBucket, ClusterRow)
- format.ts       — fmtHours / fmtPct / fmtUsd
- ImpactSections.tsx — CycleTimeSection / EngagementSection / CostSection
                     (+ their StatCard, spread, and chart helpers)

InsightsClient.tsx now imports the shared types + the three sections; the
moved code (55KB -> 35KB) is gone, the FP / accuracy views are untouched.

Part of #218.
